codes/requirements.txt
# Cataract-LMM Unified Requirements
# Generated by Senior AI Engineer - Project Audit & Refinement
# Compatible with Python 3.8+

# ============================================================================
# Core Deep Learning Framework
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=0.12.0,<2.1.0

# ============================================================================
# Computer Vision & Video Processing
# ============================================================================
opencv-python-headless>=4.6.0,<5.0.0
Pillow>=9.0.0,<11.0.0
scikit-image>=0.19.0,<0.22.0

# Video-specific libraries
pytorchvideo>=0.1.5
decord>=0.6.0
av>=8.0.0
moviepy>=1.0.3

# ============================================================================
# ML & Scientific Computing
# ============================================================================
numpy>=1.21.0,<1.25.0
pandas>=1.4.0,<2.1.0
scikit-learn>=1.1.0,<1.4.0
scipy>=1.8.0,<1.12.0

# ML Metrics and Utilities
torchmetrics>=0.9.0,<1.1.0

# ============================================================================
# Deep Learning Models & Architectures
# ============================================================================
# Object Detection & Segmentation
ultralytics>=8.0.0  # YOLOv8/YOLOv11
# detectron2>=0.6  # Mask R-CNN (uncomment if needed)

# Transformer models
timm>=0.6.0  # For various CNN/ViT backbones
transformers>=4.20.0  # For transformer-based models
einops>=0.4.0  # For tensor operations

# Video Classification models
timesformer-pytorch>=0.1.0

# Segmentation models (optional)
# segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# ============================================================================
# Data Processing & Augmentation
# ============================================================================
albumentations>=1.3.0  # Advanced image augmentations
imgaug>=0.4.0  # Image augmentation library

# ============================================================================
# Configuration & File Management
# ============================================================================
PyYAML>=6.0,<7.0
pydantic>=1.10.0,<2.0  # Data validation
click>=8.0.0  # CLI interface
pathlib2>=2.3.0  # Enhanced path handling (for Python <3.4 compatibility)

# ============================================================================
# Visualization & Plotting
# ============================================================================
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<0.13.0
plotly>=5.0.0,<6.0.0

# ============================================================================
# Progress & Logging
# ============================================================================
tqdm>=4.64.0
rich>=12.0.0  # Enhanced console output
colorlog>=6.7.0  # Colored logging

# ============================================================================
# Jupyter & Interactive Development
# ============================================================================
jupyter>=1.0.0
ipywidgets>=7.7.0,<9.0.0
nbconvert>=6.5.0
ipykernel>=6.0.0

# ============================================================================
# Development & Testing Dependencies
# ============================================================================
# Code Quality
black>=22.0.0  # Code formatter
isort>=5.10.0  # Import sorter
flake8>=4.0.0  # Linter
mypy>=0.991  # Type checker

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0  # Coverage
pytest-mock>=3.10.0  # Mocking
hypothesis>=6.0.0  # Property-based testing

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Performance Profiling
memory-profiler>=0.60.0
line-profiler>=4.0.0

# ============================================================================
# Optional: GPU Acceleration
# ============================================================================
# Uncomment these for specific CUDA versions:
# torch==1.13.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision==0.14.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html

# ============================================================================
# Optional: Additional ML Libraries
# ============================================================================
# xgboost>=1.6.0  # Gradient boosting
# lightgbm>=3.3.0  # Gradient boosting
# catboost>=1.0.0  # Gradient boosting
# mlflow>=2.0.0  # Experiment tracking
# wandb>=0.13.0  # Experiment tracking

# ============================================================================
# Compatibility & Utilities
# ============================================================================
typing-extensions>=4.0.0  # Enhanced typing support
packaging>=21.0  # Version parsing
requests>=2.28.0  # HTTP requests
urllib3>=1.26.0,<2.0.0  # URL handling

# ============================================================================
# END - Unified Requirements
# ============================================================================
